MULE-12973: Add support for start parameter in http multipart/related response (#16)
According to this RFC article: https://tools.ietf.org/html/rfc2387, in multipart/related content type: The start parameter, if given, is the content-ID
of the compound object's "root". If not present the "root" is the first body part in the Multipart/Related entity. Additionally, we shoud also honor
the following restrictions:

- Type parameter is mandatory in multipart/related.
- The value of the content-type parameter must be the same that the root body part's content-type.